refactor: move hslToRgb to utils
